pyproject.toml
[tool.poetry]
name = "netrise-turbine-sdk"
version = "0.1.13"
description = "Turbine GraphQL Python SDK (generated via ariadne-codegen)"
authors = ["NetRise <anthony.feddersen@netrise.io>"]
readme = "README.md"
packages = [
  { include = "netrise_turbine_sdk", from = "src" },
  { include = "netrise_turbine_sdk_graphql", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.12"
httpx = ">=0.27.0,<1.0.0"
pydantic = ">=2.0.0,<3.0.0"
python-dotenv = ">=1.0.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
ariadne-codegen = {version = ">=0.17.0,<0.18.0", python = "<4"}
graphql-core = {version = "==3.2.0", python = "<4"}
pytest = ">=8.0.0,<9.0.0"
pytest-cov = "^7.0.0"

[tool.ariadne-codegen]
schema_path = "../sdk-artifacts/schema.graphql"
queries_path = "../sdk-artifacts/synthetic_operations.graphql"
# Generate into sdk/python-sdk/src/netrise_turbine_sdk_graphql
# (keeps handwritten wrapper code separate)
target_package_name = "netrise_turbine_sdk_graphql"
target_package_path = "src"
async_client = false
files_to_include = ["custom_scalars.py"]

[tool.ariadne-codegen.scalars.BigInt]
type = "int"

[tool.ariadne-codegen.scalars.JSON]
type = "typing.Any"

[tool.ariadne-codegen.scalars.DateTime]
type = "datetime.datetime"
parse = ".custom_scalars.parse_datetime"
serialize = ".custom_scalars.serialize_datetime"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
